refactor(web): migrate Dashboard, Chat, and Agents pages to app router convention
- Moved `web/src/pages/DashboardPage.tsx` -> `web/src/app/dashboard/page.tsx`
- Moved `web/src/pages/ChatPage.tsx` -> `web/src/app/chat/page.tsx`
- Moved `web/src/pages/AgentsPage.tsx` -> `web/src/app/agents/page.tsx`
- Updated imports in `web/src/main.tsx` and affected test files (`ChatStream.test.tsx`, `AgentDelete.test.tsx`).
- Migrated 3 pages per instructions, maintaining behavior and limits.
